requirement.txt
a==1.0
arabic-reshaper==2.1.4
asgiref==3.5.0
asn1crypto==1.5.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==38.0.1
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0.2
django-dotenv==1.4.2
django-embed-video==1.4.7
django-environ==0.9.0
django-heroku==0.3.1
django-import-export==2.9.0
django-mathfilters==1.0.0
django-phonenumber-field==6.1.0
django-stronghold==0.4.0
django-tastypie==0.14.4
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.2.0
environ==1.0
et-xmlfile==1.1.0
filelock==3.7.1
Flask==2.0.3
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.28.5
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.3.2
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.0
matplotlib==3.5.1
MouseInfo==0.1.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.10
optional-django==0.3.0
oscrypto==1.3.0
packaging==21.3
pandas==1.3.5
passlib==1.7.4
pdfkit==1.0.0
phonenumbers==8.12.47
platformdirs==2.5.2
protobuf==3.19.3
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
PyMsgBox==1.0.9
pyparsing==3.0.6
PyPDF3==1.0.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-mimeparse==1.6.0
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
qrcode==7.3.1
razorpay==1.3.0
react==4.3.0
reportlab==3.6.11
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlparse==0.4.2
svglib==1.4.1
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.1
twilio==7.8.2
tzdata==2021.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
virtualenv==20.14.1
waitress==2.1.2
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==6.2.0
wikipedia==1.4.0
WTForms==3.0.1
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
